requirements.txt
pandas>=2.0.0
pyarrow>=12.0.0
pyspark==3.5.0
great-expectations>=0.18.0
thefuzz>=0.19.0
python-Levenshtein>=0.21.0
libpostal>=1.1.0
spacy>=3.6.0
scikit-learn>=1.3.0
xgboost>=2.0.0
prefect>=2.14.0
streamlit>=1.28.0
matplotlib>=3.7.0
seaborn>=0.12.0
phonenumbers>=8.13.0
email-validator>=2.0.0
usaddress>=0.5.10
pyyaml>=6.0.1
boto3>=1.28.0
numpy>=1.24.0
jellyfish>=0.11.0
click>=8.1.0
pytest>=7.4.0
pytest-cov>=4.1.0
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
